Fix `into sqlite`
Currently the `into sqlite` command collects an entire input stream into
a single `Value` before doing any writes. This leads to massive memory
consumption if the input size is large, as the entire input is loaded
into memory.

This changes the command to work directly with the `PipelineData` rather
than collect first.