from csv: stream
Fixes memory consumption of `from csv` by changing it to steam its input
and output, rather than soak up everything into memory.

This is accomplished with a new type `StreamReader` that implements
`std::io::Read` to give direct access to the bytes of the underlying
values, of the values are a type that can give back bytes in a
semantically unambiguous way. Types that cannot yield an error in the
reads.

This reader is used as the input source of the `csv::Reader` so that it
can be properly steamed.